refactored producers page to use the same sidebar between block producers and proxies
